deprecated/v2/requirements.txt
elevenlabs==1.58.1
gaveta==0.6.0
kokoro==0.9.4
mypy==1.12.0
python-dotenv==1.1.0
ruff==0.11.2
scipy-stubs[scipy]==1.15.3.0
transformers[torch]==4.51.3
